Fixes quadArc rendering problem
