+++ lib/scripts/effects.js
@@ -38,7 +38,7 @@ Effects.prototype.introSequence = function(self) {
     $('.level-announcements').fadeOut(1000);
     $('canvas').delay(2000).fadeIn(1000);
     $('.game-div').off('click');
-    self.playLevel(9);
+    self.playLevel(5);
   });
 };
