+++ lib/scripts/renderer.js
@@ -109,7 +109,7 @@ Renderer.prototype.renderQuadArc = function(location) {
   this.context.clearRect(startX, startY, 40, 40);
   this.context.strokeStyle = "#BFAF80";
   this.context.beginPath();
-  this.context.arc(startX, startY, 20, (Math.PI * 1.5), Math.PI, false);
+  this.context.arc(startX, startY, 20, 0, (Math.PI * 0.5), false);
   this.context.stroke();
   this.context.beginPath();
   this.context.arc(startX + 40, startY, 20, (Math.PI * 0.5), Math.PI, false);
@@ -142,11 +142,11 @@ let lineCoordinates = {
 };
 
 Renderer.prototype.findStartX = function(location) {
-  return this.grid.locationsIndex[location].x;
+  return this.grid.tiles.tileIndex([location]).x;
 };
 
 Renderer.prototype.findStartY = function(location) {
-  return this.grid.locationsIndex[location].y;
+  return this.grid.tiles.tileIndex([location]).y;
 };
 
 module.exports = Renderer;
