+++ lib/scripts/tiles.js
@@ -1,0 +1,135 @@
+let Tiles = function() {};
+
+Tiles.prototype.tileNumber = function(x, y) {
+  let tilesByCoordinate = {
+    60:  { 20: 'A1', 60: 'A2', 100: 'A3', 140: 'A4', 180: 'A5', 220: 'A6', 260: 'A7', 300: 'A8', 340: 'A9' },
+    100: { 20: 'B1', 60: 'B2', 100: 'B3', 140: 'B4', 180: 'B5', 220: 'B6', 260: 'B7', 300: 'B8', 340: 'B9' },
+    140: { 20: 'C1', 60: 'C2', 100: 'C3', 140: 'C4', 180: 'C5', 220: 'C6', 260: 'C7', 300: 'C8', 340: 'C9' },
+    180: { 20: 'D1', 60: 'D2', 100: 'D3', 140: 'D4', 180: 'D5', 220: 'D6', 260: 'D7', 300: 'D8', 340: 'D9' },
+    220: { 20: 'E1', 60: 'E2', 100: 'E3', 140: 'E4', 180: 'E5', 220: 'E6', 260: 'E7', 300: 'E8', 340: 'E9' },
+    260: { 20: 'F1', 60: 'F2', 100: 'F3', 140: 'F4', 180: 'F5', 220: 'F6', 260: 'F7', 300: 'F8', 340: 'F9' },
+    300: { 20: 'G1', 60: 'G2', 100: 'G3', 140: 'G4', 180: 'G5', 220: 'G6', 260: 'G7', 300: 'G8', 340: 'G9' },
+    340: { 20: 'H1', 60: 'H2', 100: 'H3', 140: 'H4', 180: 'H5', 220: 'H6', 260: 'H7', 300: 'H8', 340: 'H9' },
+    380: { 20: 'I1', 60: 'I2', 100: 'I3', 140: 'I4', 180: 'I5', 220: 'I6', 260: 'I7', 300: 'I8', 340: 'I9' },
+    420: { 20: 'J1', 60: 'J2', 100: 'J3', 140: 'J4', 180: 'J5', 220: 'J6', 260: 'J7', 300: 'J8', 340: 'J9' },
+    460: { 20: 'K1', 60: 'K2', 100: 'K3', 140: 'K4', 180: 'K5', 220: 'K6', 260: 'K7', 300: 'K8', 340: 'K9' },
+    500: { 20: 'L1', 60: 'L2', 100: 'L3', 140: 'L4', 180: 'L5', 220: 'L6', 260: 'L7', 300: 'L8', 340: 'L9' }
+  };
+  return tilesByCoordinate[x][y];
+};
+
+Tiles.prototype.tileIndex = function(index) {
+  let locationsIndex = {
+    'A1': { 'x': 60, 'y': 20},
+    'A2': { 'x': 60, 'y': 60},
+    'A3': { 'x': 60, 'y': 100},
+    'A4': { 'x': 60, 'y': 140},
+    'A5': { 'x': 60, 'y': 180},
+    'A6': { 'x': 60, 'y': 220},
+    'A7': { 'x': 60, 'y': 260},
+    'A8': { 'x': 60, 'y': 300},
+    'A9': { 'x': 60, 'y': 340},
+    'B1': { 'x': 100, 'y': 20},
+    'B2': { 'x': 100, 'y': 60},
+    'B3': { 'x': 100, 'y': 100},
+    'B4': { 'x': 100, 'y': 140},
+    'B5': { 'x': 100, 'y': 180},
+    'B6': { 'x': 100, 'y': 220},
+    'B7': { 'x': 100, 'y': 260},
+    'B8': { 'x': 100, 'y': 300},
+    'B9': { 'x': 100, 'y': 340},
+    'C1': { 'x': 140, 'y': 20},
+    'C2': { 'x': 140, 'y': 60},
+    'C3': { 'x': 140, 'y': 100},
+    'C4': { 'x': 140, 'y': 140},
+    'C5': { 'x': 140, 'y': 180},
+    'C6': { 'x': 140, 'y': 220},
+    'C7': { 'x': 140, 'y': 260},
+    'C8': { 'x': 140, 'y': 300},
+    'C9': { 'x': 140, 'y': 340},
+    'D1': { 'x': 180, 'y': 20},
+    'D2': { 'x': 180, 'y': 60},
+    'D3': { 'x': 180, 'y': 100},
+    'D4': { 'x': 180, 'y': 140},
+    'D5': { 'x': 180, 'y': 180},
+    'D6': { 'x': 180, 'y': 220},
+    'D7': { 'x': 180, 'y': 260},
+    'D8': { 'x': 180, 'y': 300},
+    'D9': { 'x': 180, 'y': 340},
+    'E1': { 'x': 220, 'y': 20},
+    'E2': { 'x': 220, 'y': 60},
+    'E3': { 'x': 220, 'y': 100},
+    'E4': { 'x': 220, 'y': 140},
+    'E5': { 'x': 220, 'y': 180},
+    'E6': { 'x': 220, 'y': 220},
+    'E7': { 'x': 220, 'y': 260},
+    'E8': { 'x': 220, 'y': 300},
+    'E9': { 'x': 220, 'y': 340},
+    'F1': { 'x': 260, 'y': 20},
+    'F2': { 'x': 260, 'y': 60},
+    'F3': { 'x': 260, 'y': 100},
+    'F4': { 'x': 260, 'y': 140},
+    'F5': { 'x': 260, 'y': 180},
+    'F6': { 'x': 260, 'y': 220},
+    'F7': { 'x': 260, 'y': 260},
+    'F8': { 'x': 260, 'y': 300},
+    'F9': { 'x': 260, 'y': 340},
+    'G1': { 'x': 300, 'y': 20},
+    'G2': { 'x': 300, 'y': 60},
+    'G3': { 'x': 300, 'y': 100},
+    'G4': { 'x': 300, 'y': 140},
+    'G5': { 'x': 300, 'y': 180},
+    'G6': { 'x': 300, 'y': 220},
+    'G7': { 'x': 300, 'y': 260},
+    'G8': { 'x': 300, 'y': 300},
+    'G9': { 'x': 300, 'y': 340},
+    'H1': { 'x': 340, 'y': 20},
+    'H2': { 'x': 340, 'y': 60},
+    'H3': { 'x': 340, 'y': 100},
+    'H4': { 'x': 340, 'y': 140},
+    'H5': { 'x': 340, 'y': 180},
+    'H6': { 'x': 340, 'y': 220},
+    'H7': { 'x': 340, 'y': 260},
+    'H8': { 'x': 340, 'y': 300},
+    'H9': { 'x': 340, 'y': 340},
+    'I1': { 'x': 380, 'y': 20},
+    'I2': { 'x': 380, 'y': 60},
+    'I3': { 'x': 380, 'y': 100},
+    'I4': { 'x': 380, 'y': 140},
+    'I5': { 'x': 380, 'y': 180},
+    'I6': { 'x': 380, 'y': 220},
+    'I7': { 'x': 380, 'y': 260},
+    'I8': { 'x': 380, 'y': 300},
+    'I9': { 'x': 380, 'y': 340},
+    'J1': { 'x': 420, 'y': 20},
+    'J2': { 'x': 420, 'y': 60},
+    'J3': { 'x': 420, 'y': 100},
+    'J4': { 'x': 420, 'y': 140},
+    'J5': { 'x': 420, 'y': 180},
+    'J6': { 'x': 420, 'y': 220},
+    'J7': { 'x': 420, 'y': 260},
+    'J8': { 'x': 420, 'y': 300},
+    'J9': { 'x': 420, 'y': 340},
+    'K1': { 'x': 460, 'y': 20},
+    'K2': { 'x': 460, 'y': 60},
+    'K3': { 'x': 460, 'y': 100},
+    'K4': { 'x': 460, 'y': 140},
+    'K5': { 'x': 460, 'y': 180},
+    'K6': { 'x': 460, 'y': 220},
+    'K7': { 'x': 460, 'y': 260},
+    'K8': { 'x': 460, 'y': 300},
+    'K9': { 'x': 460, 'y': 340},
+    'L1': { 'x': 500, 'y': 20},
+    'L2': { 'x': 500, 'y': 60},
+    'L3': { 'x': 500, 'y': 100},
+    'L4': { 'x': 500, 'y': 140},
+    'L5': { 'x': 500, 'y': 180},
+    'L6': { 'x': 500, 'y': 220},
+    'L7': { 'x': 500, 'y': 260},
+    'L8': { 'x': 500, 'y': 300},
+    'L9': { 'x': 500, 'y': 340}
+  };
+  return locationsIndex[index];
+};
+
+module.exports = Tiles;
